requirements.txt
pyperclip
llama-cpp-python